// src/app/models/ICustomer2.ts
export interface ICustomer {
    id:string
    firstName: string
    lastName: string
    email: string
    password: string
    userName: string
    doB: string
    phoneNumber: number
    parentId: string
    nominee: string
    nomineeRelation: string
    city: string
    state: string
    address: string
    pincode:number
}